Cleaned up article worker.  Moved article model and db interaction to server/article model and controller files.  Got rid of some legacy stuff (eg feedCat)
